Database heroku fix

+++ server.js
@@ -28,8 +28,14 @@ require("./routes/html-routes.js")(app);
 
 // Syncing our sequelize models and then starting our Express app
 // =============================================================
-db.sequelize.sync({ force: true }).then(function() {
+if (process.env.ENV === "PRODUCTION") {
   app.listen(PORT, function() {
     console.log("App listening on PORT " + PORT);
   });
-});
+} else {
+  db.sequelize.sync({ force: true }).then(function() {
+    app.listen(PORT, function() {
+      console.log("App listening on PORT " + PORT);
+    });
+  });
+}
